@@ -14,9 +14,11 @@
         }
     ],
     "require": {
-        "php": ">=5.6.0",
+        "php": ">=7.4.0",
         "ext-mcrypt": "*",
-        "ext-openssl": "*"
+        "ext-openssl": "*",
+        "ext-json": "*",
+        "ext-curl": "*"
     },
     "autoload": {
         "psr-4": {
