@@ -13,22 +13,19 @@ namespace iRAP\VidaSDK\Models;
 
 class AppAuthentication extends AbstractAuthentication
 {
-    protected $m_app_auth_id;
-    protected $m_app_api_key;
-    protected $m_app_private_key;
-    
-    
+    protected int $m_app_auth_id;
+    protected string $m_app_api_key;
+    protected string $m_app_private_key;
+
+
     /**
      * Takes the API token and sets up the authentication member variable
-     * 
+     *
      * @param int $app_auth_id
      * @param string $app_api_key
      * @param string $app_private_key
-     * @param int $user_auth_id
-     * @param string $user_api_key
-     * @param string $user_private_key
      */
-    public function __construct($app_auth_id, $app_api_key, $app_private_key)
+    public function __construct(int $app_auth_id, string $app_api_key, string $app_private_key)
     {
         $this->m_app_auth_id = $app_auth_id;
         $this->m_app_api_key = $app_api_key;
@@ -41,7 +38,7 @@ class AppAuthentication extends AbstractAuthentication
      * @param string $message
      * @return string
      */
-    public function getEncryption($message)
+    public function getEncryption(string $message): string
     {
         return $this->encrypt($message, $this->m_app_private_key);
     }
@@ -51,22 +48,21 @@ class AppAuthentication extends AbstractAuthentication
      * 
      * @return array
      */
-    public function getAuthHeaders()
+    public function getAuthHeaders(): array
     {
-        $parameters = array(
+        return array(
             'auth_system_auth_id'    => $this->m_app_auth_id,
             'auth_system_public_key' => $this->m_app_api_key
         );
-        
-        return $parameters;
     }
-    
+
     /**
      * Gets the signature for app and returns as an array.
-     * 
+     *
+     * @param array $data
      * @return array
      */
-    public function getSignatures(array $data)
+    public function getSignatures(array $data): array
     {
         return array(
             'auth_system_signature'  => $this->generateSignature($data, $this->m_app_private_key)
