@@ -13,14 +13,14 @@ namespace iRAP\VidaSDK\Models;
 
 class UserAuthentication extends AbstractAuthentication
 {
-    protected $m_app_auth_id;
-    protected $m_app_api_key;
+    protected int $m_app_auth_id;
+    protected string $m_app_api_key;
     protected $m_app_protected_key;
-    protected $m_user_auth_id;
-    protected $m_user_api_key;
+    protected int $m_user_auth_id;
+    protected string $m_user_api_key;
     protected $m_user_protected_key;
-    protected $m_app_private_key;
-    protected $m_user_private_key;
+    protected string $m_app_private_key;
+    protected string $m_user_private_key;
     
     /**
      * Takes the API token and user token if available and sets up the authentication member variable
@@ -32,7 +32,7 @@ class UserAuthentication extends AbstractAuthentication
      * @param string $user_api_key
      * @param string $user_private_key
      */
-    public function __construct($app_auth_id, $app_api_key, $app_private_key, $user_auth_id, $user_api_key, $user_private_key)
+    public function __construct(int $app_auth_id, string $app_api_key, string $app_private_key, int $user_auth_id, string $user_api_key, string $user_private_key)
     {
         $this->m_app_auth_id = $app_auth_id;
         $this->m_app_api_key = $app_api_key;
@@ -48,25 +48,23 @@ class UserAuthentication extends AbstractAuthentication
      * 
      * @return array
      */
-    public function getAuthHeaders()
+    public function getAuthHeaders(): array
     {
-        $parameters = array(
+        return array(
             'auth_system_auth_id'    => $this->m_app_auth_id,
             'auth_system_public_key' => $this->m_app_api_key,
             'auth_user_auth_id'      => $this->m_user_auth_id,
             'auth_user_public_key'   => $this->m_user_api_key,
         );
-        
-        return $parameters;
     }
     
     
     /**
      * Gets the user and app signatures for the provided data 
-     * before returning them as an assosciative array.
+     * before returning them as an associative array.
      * @return array - name/value pairs of the signatures.
      */
-    public function getSignatures(array $data)
+    public function getSignatures(array $data): array
     {
         return array(
             'auth_system_signature'  => $this->generateSignature($data, $this->m_app_private_key),
