@@ -6,24 +6,29 @@
 
 namespace iRAP\VidaSDK;
 
-Interface FilterInterface extends \JsonSerializable
+use JsonSerializable;
+
+Interface FilterInterface extends JsonSerializable
 {
     /**
      * The same as getFilter, but performs a urlencode on the result.
      * @return string
      */
-    public function buildFilter();
+    public function buildFilter(): string;
     
     
     /**
      * Convert this object into json form for the API
      * @return string - json
      */
-    public function getFilter();
+    public function getFilter(): string;
     
     
     /**
      * All filters need to be serializable
+     * @return mixed
+     * @TODO Attribute ReturnTypeWillChange requires updating once PHP requirement for this package is at least 8
      */
-    public function jsonSerialize(): mixed ;
+    #[\ReturnTypeWillChange]
+    public function jsonSerialize();
 }
