@@ -11,9 +11,14 @@
 
 namespace iRAP\VidaSDK\Controllers;
 
+use Exception;
+use iRAP\VidaSDK\Defines;
+use iRAP\VidaSDK\Models\AbstractAuthentication;
+use iRAP\VidaSDK\Models\APIRequest;
+
 class AuthController extends AbstractResourceController
 {
-    protected function getResourcePath()
+    protected function getResourcePath(): string
     {
         return 'auth';
     }
@@ -23,14 +28,16 @@ class AuthController extends AbstractResourceController
      * should be stored locally and used for all future requests. Email and password should NOT
      * be stored locally. The password is encrypted using the APP_PRIVATE KEY before transmission.
      *
+     * @param AbstractAuthentication $auth
      * @param string $email
      * @param string $password
      * @return object
+     * @throws Exception
      */
-    public static function getUserToken(\iRAP\VidaSDK\Models\AbstractAuthentication $auth, $email, $password)
+    public static function getUserToken(AbstractAuthentication $auth, string $email, string $password): object
     {
         $encrypted_password = $auth->getEncryption($password);
-        $request = new \iRAP\VidaSDK\Models\APIRequest($auth);
+        $request = new APIRequest($auth);
         $request->setUrl("auth/register");
         $request->setPostData(array("email"=>$email,"password"=>$encrypted_password));
         $request->send();
@@ -41,10 +48,10 @@ class AuthController extends AbstractResourceController
      * Builds query parameters to sign, signs them and then sends the query to ViDA, so that the
      * user can view and accept/reject the permissions that the app is asking for.
      *
-     * @param \iRAP\VidaSDK\Models\AbstractAuthentication $auth
-     * @param type $returnUrl
+     * @param AbstractAuthentication $auth
+     * @param string $returnUrl
      */
-    public static function requestUserPermissions(\iRAP\VidaSDK\Models\AbstractAuthentication $auth, $returnUrl)
+    public static function requestUserPermissions(AbstractAuthentication $auth, string $returnUrl)
     {
         $headers = $auth->getAuthHeaders();
         $headers['return_url'] = urlencode($returnUrl);
@@ -57,7 +64,7 @@ class AuthController extends AbstractResourceController
         }
         else
         {
-            $url = \iRAP\VidaSDK\Defines::IRAP_PERMISSIONS_LIVE_URL;
+            $url = Defines::IRAP_PERMISSIONS_LIVE_URL;
         }
 
         header('Location: ' . $url . '?' . http_build_query($query));
