@@ -12,7 +12,7 @@ namespace iRAP\VidaSDK\Controllers;
 
 class DatasetsController extends AbstractResourceController
 {
-    protected function getResourcePath()
+    protected function getResourcePath(): string
     {
         return 'datasets';
     }
