@@ -4,7 +4,7 @@
  * This file is for internal use by the ViDA SDK. It should not be altered by users
  **************************************************************************************************
  * 
- * Controller for the Star Ratings Restuls Summary resource. Overrides the Abstract Resource 
+ * Controller for the Star Ratings Results Summary resource. Overrides the Abstract Resource
  * Controller.
  * 
  */
@@ -13,7 +13,7 @@ namespace iRAP\VidaSDK\Controllers;
 
 class StarRatingResultsSummaryController extends AbstractResourceController
 {
-    protected function getResourcePath()
+    protected function getResourcePath(): string
     {
         return 'star-rating-results-summary';
     }
